Разделение классов и интерфейсов на собственные файлы. Упрощение и разбиение кода парсера.
