+++ src/HRInPocket/Services/HRInPocket.Parsing.hh.ru/HRInPocket.Parsing.hh.ru/Interfaces/IParsehh.cs
@@ -1,18 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-using HRInPocket.Parsing.hh.ru.Models.Entites;
 using HRInPocket.Parsing.hh.ru.Service;
 
 namespace HRInPocket.Parsing.hh.ru.Interfaces
 {
-    public interface IParsehhService
-    {
-        IParsehh GetPasrse();
-    }
     public interface IParsehh
     {
         event EventHandler<VacancyEventArgs> Result;
