+++ src/HRInPocket/Services/HRInPocket.Parsing.hh.ru/HRInPocket.Parsing.hh.ru/Interfaces/IParsehhService.cs
@@ -1,0 +1,7 @@
+﻿namespace HRInPocket.Parsing.hh.ru.Interfaces
+{
+    public interface IParsehhService
+    {
+        IParsehh GetPasrse();
+    }
+}
