+++ src/HRInPocket/Services/HRInPocket.Parsing.hh.ru/HRInPocket.Parsing.hh.ru/Models/Entites/Vacancy.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data.SqlTypes;
-using System.Text;
 
 using HRInPocket.Parsing.hh.ru.Models.Base;
