+++ src/HRInPocket/Services/HRInPocket.Parsing.hh.ru/HRInPocket.Parsing.hh.ru/Models/Entites/VacancyName.cs
@@ -1,13 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-using HRInPocket.Parsing.hh.ru.Models.Base;
+﻿using HRInPocket.Parsing.hh.ru.Models.Base;
 
 namespace HRInPocket.Parsing.hh.ru.Models.Entites
 {
-    public class VacancyName: BaseUnit
-    {
-       
-    }
+    public class VacancyName : BaseUnit { }
 }
