+++ src/HRInPocket/Services/HRInPocket.Parsing.hh.ru/HRInPocket.Parsing.hh.ru/Service/VacancyEventArgs.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+using HRInPocket.Parsing.hh.ru.Models.Entites;
+
+namespace HRInPocket.Parsing.hh.ru.Service
+{
+    public class VacancyEventArgs : EventArgs
+    {
+        public Vacancy Vacancy { get; set; }
+    }
+}
+
